experiement with Progress. Still appears to be waiting until app terminates before printing.

+++ lib/script/project.dart
@@ -107,7 +107,9 @@ class VirtualProject {
   ///
   /// deletes the project cache directory and recreates it.
   void clean() {
-    File(_projectRootPath).deleteSync(recursive: true);
+    if (exists(_projectRootPath)) {
+      File(_projectRootPath).deleteSync(recursive: true);
+    }
 
     createProject();
   }
